requirements.txt
numpy
pillow
openpyxl
pandas